[WIP] Add function to extract base specifiers
Certain base specifiers may not appear in the clang abstract syntax tree
(AST).  This seems to occur for base specifiers that are class templates
where the template types cannot be resolved.

As a workaround, the `extract_base_identifiers` extracts a list of all
`IDENTIFIER` tokens found in the base specifier list of a class.  Class
names found in the identifier list but not in the AST may, for example,
be added to the AST.