feat(c++): add clang cxx headers to includes
Prior to this commit, if a C++ source file was opened using
`open_cpp_source()`, parsing would fail for any code involving the C++
standard library due to missing headers unless a path to C++ STL
headerswas specified, either explicitly as arguments to the function or
through the `CPATH` environment variable.

In this commit, require the new `clang-libcxx` Conda package and add the
header directory from the `clang-libcxx` Conda package to the includes
path in `open_cpp_source()` (Windows only).